* Move tests into proper testing framework.
