+++ tests/test_radio.py
@@ -1,46 +1,7 @@
-#!/usr/bin/env python3
-
-import time
 import unittest
 
 from sts.client import Datum, Radio
-
-
-class DatumTest(unittest.TestCase):
-    """Unit tests for the Datum class."""
-
-    @staticmethod
-    def create_data():
-        """Create a list of STS data that can be written to STS for testing."""
-
-        # Long text data (62*3 bytes)
-        text = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz' * 3
-        timestamp = int(time.time())
-        data = [
-            Datum.Integer(id=1090, timestamp=timestamp, value=1),
-            Datum.Float(id=1091, timestamp=timestamp, value=1.0),
-            Datum.Text(id=1092, timestamp=timestamp, value=text),
-            Datum.IntegerWithText(id=1093, timestamp=timestamp, value=(1, text)),
-            Datum.FloatWithText(id=1094, timestamp=timestamp, value=(1.0, text)),
-            Datum.Exponent(id=1095, timestamp=timestamp, value=1.0),
-        ]
-        return data
-
-    @staticmethod
-    def create_ids():
-        """Create a list of STS radio IDs that can be read from STS for testing."""
-
-        return [1090, 1091, 1092, 1093, 1094, 1095]
-
-    def test_default_constructor(self):
-        """Test the default constructor of the Datum class."""
-
-        datum = Datum()
-
-    def test_factory_methods(self):
-        """Test the factory class methods of the Datum class."""
-
-        DatumTest.create_data()
+from .test_datum import DatumTest
 
 
 class RadioTest(unittest.TestCase):
@@ -49,7 +10,7 @@ class RadioTest(unittest.TestCase):
     def test_default_constructor(self):
         """Test the default constructor of the Radio class."""
 
-        radio = Radio()
+        _ = Radio()
 
     def test_pack_method_with_invalid_data_type(self):
         """Test the pack static method with an invalid data type."""
@@ -93,9 +54,4 @@ class RadioTest(unittest.TestCase):
 
         radio = Radio()
         ids = DatumTest.create_ids()
-        data = radio.receive(ids)
-
-
-if __name__ == '__main__':
-
-    unittest.main()
+        _ = radio.receive(ids)
